requirements.txt
asgiref==3.5.0
atomicwrites==1.4.0
attrs==19.3.0
beautifulsoup4==4.10.0
certifi==2019.9.11
cffi==1.15.0
chardet==3.0.4
charset-normalizer==2.0.11
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==36.0.2
defusedxml==0.7.1
Django==2.2.16
django-debug-toolbar==2.2
django-filter==1.1.0
django-templated-mail==1.1.1
djangorestframework==3.13.1
djangorestframework-filters==0.11.1
djangorestframework-simplejwt==4.7.2
djoser==2.1.0
Faker==12.0.1
flake8==4.0.1
idna==2.8
importlib-metadata==1.5.0
iniconfig==1.1.1
isort==5.10.1
itypes==1.2.0
Jinja2==3.0.3
MarkupSafe==2.1.1
mccabe==0.6.1
mixer==7.1.2
more-itertools==8.2.0
oauthlib==3.2.0
packaging==20.1
Pillow==8.3.1
pluggy==0.13.1
py==1.11.0
pycodestyle==2.8.0
pycparser==2.21
pyflakes==2.4.0
PyJWT==2.3.0
pyparsing==2.4.6
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2019.3
requests==2.26.0
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.2.0
sorl-thumbnail==12.7.0
soupsieve==2.3.1
sqlparse==0.3.0
text-unidecode==1.3
toml==0.10.2
typing_extensions==4.0.1
uritemplate==4.1.1
urllib3==1.25.6
wcwidth==0.1.8
zipp==2.2.0
